+++ profiler/src/ProfilerEngine/Datadog.Profiler.Native/FrameStore.cpp
@@ -7,8 +7,8 @@
 #include "OpSysTools.h"
 
 #include "shared/src/native-src/com_ptr.h"
-#include "shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 
 FrameStore::FrameStore(ICorProfilerInfo4* pCorProfilerInfo) :
     _pCorProfilerInfo{pCorProfilerInfo}
