+++ profiler/src/ProfilerEngine/Datadog.Profiler.Native/IConfiguration.h
@@ -9,8 +9,8 @@
 
 #include "TagsHelper.h"
 
-#include "shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 #include "shared/src/native-src/string.h"
 
 class IConfiguration
