+++ profiler/src/ProfilerEngine/Datadog.Profiler.Native/LibddprofExporter.cpp
@@ -14,8 +14,8 @@
 #include <string.h>
 #include <time.h>
 
-#include "shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 
 #define BUFFER_MAX_SIZE 512
