+++ profiler/src/ProfilerEngine/Datadog.Profiler.Native/OpSysTools.cpp
@@ -26,8 +26,8 @@
 #include <chrono>
 #include <thread>
 
-#include "shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 
 #include "Log.h"
