+++ profiler/test/Datadog.Profiler.Native.Tests/ProfilerMockedInterface.h
@@ -2,8 +2,8 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
-#include "shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 
 #include "Configuration.h"
 #include "IExporter.h"
