+++ tracer/src/Datadog.AutoInstrumentation.NativeLoader/dynamic_dispatcher.cpp
@@ -4,8 +4,8 @@
 #include <unordered_map>
 
 #include "log.h"
-#include "../../../shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "../../../shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 #include "../../../shared/src/native-src/pal.h"
 #include "../../../shared/src/native-src/util.h"
