+++ tracer/src/Datadog.AutoInstrumentation.NativeLoader/dynamic_dispatcher.h
@@ -7,8 +7,8 @@
 
 #include "dynamic_instance.h"
 
-#include "../../../shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "../../../shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 #include "../../../shared/src/native-src/string.h"
 
 namespace datadog::shared::nativeloader
