+++ tracer/src/Datadog.AutoInstrumentation.NativeLoader/util.h
@@ -9,8 +9,8 @@ EXTERN_C IMAGE_DOS_HEADER __ImageBase;
 #include <dlfcn.h>
 #endif
 
-#include "../../../shared/src/native-src/filesystem.h"
-// namespace fs is an alias defined in "filesystem.h"
+#include "../../../shared/src/native-src/dd_filesystem.hpp"
+// namespace fs is an alias defined in "dd_filesystem.hpp"
 #include "../../../shared/src/native-src/pal.h"
 #include "../../../shared/src/native-src/string.h"
